pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloud-gym"
version = "0.1.0"
description = "Scalable Training Data Generation for Infrastructure-as-Code Repair via Environment Inversion"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Jon" }]
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-hcl2>=5.0",
    "cfn-lint>=1.0",
    "rich>=13.0",
    "datasets>=3.0",
    "ollama>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
cloud-gym = "cloudgym.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["cloudgym"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
